Fix added integer u32
```
xevdm_eco.c:2212:18: warning: comparison of integer expressions of different signedness: 'u32' {aka 'unsigned int'} and 'int' [-Wsign-compare]
 2212 |     if(ruiSymbol >= uiVal - b)
      |                  ^~
```

+++ src_main/xevdm_eco.c
@@ -2209,7 +2209,7 @@ u32 xevdm_xReadTruncBinCode(XEVD_BSR * bs, const int uiMaxSymbol)
     int uiVal = 1 << uiThresh;
     int b = uiMaxSymbol - uiVal;
     xevd_bsr_read(bs, &ruiSymbol, uiThresh); //xReadCode( uiThresh, ruiSymbol );
-    if(ruiSymbol >= uiVal - b)
+    if(ruiSymbol >= u32(uiVal - b))
     {
         u32 uiSymbol;
         xevd_bsr_read1(bs, &uiSymbol); //xReadFlag( uiSymbol );
